Updated the model

+++ bedriftshelsetjenesten/src/BedriftshelsetjenesteGodkjenningsregister/Avdeling.java
@@ -2,6 +2,7 @@
  */
 package BedriftshelsetjenesteGodkjenningsregister;
 
+import java.math.BigDecimal;
 import org.eclipse.emf.ecore.EObject;
 
 /**
@@ -73,12 +74,12 @@ public interface Avdeling extends EObject {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Telefon</em>' attribute.
-	 * @see #setTelefon(int)
+	 * @see #setTelefon(BigDecimal)
 	 * @see BedriftshelsetjenesteGodkjenningsregister.BedriftshelsetjenestenPackage#getAvdeling_Telefon()
-	 * @model
+	 * @model dataType="BedriftshelsetjenesteGodkjenningsregister.Telefon8Siffer"
 	 * @generated
 	 */
-	int getTelefon();
+	BigDecimal getTelefon();
 
 	/**
 	 * Sets the value of the '{@link BedriftshelsetjenesteGodkjenningsregister.Avdeling#getTelefon <em>Telefon</em>}' attribute.
@@ -88,7 +89,7 @@ public interface Avdeling extends EObject {
 	 * @see #getTelefon()
 	 * @generated
 	 */
-	void setTelefon(int value);
+	void setTelefon(BigDecimal value);
 
 	/**
 	 * Returns the value of the '<em><b>Postadresse</b></em>' reference.
